requirements.txt
# ─────────────────────────────────────────────────────────────
#  xais-brain — Dépendances Python
# ─────────────────────────────────────────────────────────────

# Config
python-dotenv>=1.0.0

# Extraction de fichiers
pypdf>=5.0.0
python-docx>=1.1.0

# LLM providers (les 3 installés, choisis via LLM_PROVIDER dans .env)
google-genai>=0.3.0
anthropic>=0.40.0
openai>=1.50.0
